requirements.txt
pyserial
setuptools